pyproject.toml
[project]
name = "sfbt-bot"
version = "0.1.0"
description = "SFBT Telegram bot"
authors = [
    {name = "maksim12", email = "maksi.volodin.05@mail.ru"}
]
requires-python = ">=3.12,<3.14"
dependencies = [
    "aiogram (>=3.25.0,<4.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic (>=2.10.0,<3.0.0)",
    "pydantic-settings (>=2.7.0,<3.0.0)",
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "redis (>=7.2.1,<8.0.0)",
    "asyncpg (>=0.30.0,<1.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "aiohttp-socks (>=0.11.0,<0.12.0)",
    "aio-pika (>=9.6.2,<10.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"